Break down resolve.go
